@@ -1,13 +1,24 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import './index.css'
-import App from './App'
-import registerServiceWorker from './registerServiceWorker'
+import React from 'react';
+import { createRoot } from 'react-dom/client'; // Importez createRoot
+import './index.css';
+import App from './App';
+import registerServiceWorker from './registerServiceWorker';
+import ErrorBoundary from './ErrorBoundary';
+// Importation des modules pour la mise en évidence du code
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import jsx from 'react-syntax-highlighter/dist/esm/languages/prism/jsx';
+import { prism } from 'react-syntax-highlighter/dist/esm/styles/prism';
 
-import { registerLanguage } from 'react-syntax-highlighter/prism-light'
-import jsx from 'react-syntax-highlighter/languages/prism/jsx'
+// Sélectionnez le conteneur DOM
+const container = document.getElementById('demo');
 
-registerLanguage('jsx', jsx)
+// Vérifiez si le conteneur DOM existe
+if (!container) {
+  throw new Error("Target container 'demo' is not a DOM element.");
+}
 
-ReactDOM.render(<App />, document.getElementById('demo'))
-registerServiceWorker()
+// Créez la racine et rendez l'application
+const root = createRoot(container);
+root.render(<ErrorBoundary><App /></ErrorBoundary>);
+
+registerServiceWorker();
