@@ -1,5 +1,5 @@
-import styled, { css } from 'styled-components'
-import { Link } from 'react-router-dom'
+import styled, { css } from 'styled-components';
+import { Link } from 'react-router-dom';
 
 const styles = css`
   background: #24B3C8;
@@ -12,25 +12,28 @@ const styles = css`
   outline: 0;
   margin: 4px;
   padding: 8px 16px;
-  text-shadow: 0 1px 0 rgba(0,0,0,0.1);
+  text-shadow: 0 1px 0 rgba(0, 0, 0, 0.1);
   text-transform: uppercase;
 
   transition: background 0.21s ease-in-out;
 
-  &:focus, &:hover {
+  &:focus,
+  &:hover {
     background: #176572;
   }
 
-  ${props => props.active && `
+  ${({ $active }) =>
+    $active &&
+    `
     background: #176572;
   `}
-`
+`;
 
 export default styled.button`
-  ${props => styles}
-`
+  ${styles}
+`;
 
 export const ButtonLink = styled(Link)`
   text-decoration: none;
-  ${props => styles}
-`
+  ${styles}
+`;
