@@ -1,9 +1,19 @@
-import React from 'react'
-import SyntaxHighlighter from 'react-syntax-highlighter/prism-light'
-import prism from 'react-syntax-highlighter/styles/prism/prism'
+import React from 'react';
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { prism } from 'react-syntax-highlighter/dist/esm/styles/prism';
 
-export default (props) => (
-  <SyntaxHighlighter language='jsx' style={prism}>
-    {props.children}
-  </SyntaxHighlighter>  
-)
+const CodeHighlighter = ({ children }) => {
+  if (!children || typeof children !== 'string') {
+    console.error('Invalid content passed to SyntaxHighlighter');
+    return null;
+  }
+
+  return (
+    <SyntaxHighlighter language="jsx" style={prism}>
+      {children}
+    </SyntaxHighlighter>
+  );
+};
+
+export default CodeHighlighter;
+
