@@ -62,20 +62,20 @@ export default class Multiple extends Component {
         <div>
           <Button
             onClick={this.onChangeType}
-            active={RectangleSelector.TYPE === this.state.type}
+            $active={RectangleSelector.TYPE === this.state.type}
           >
             {RectangleSelector.TYPE}
           </Button>
           <Button
             onClick={this.onChangeType}
-            active={PointSelector.TYPE === this.state.type}
+            $active={PointSelector.TYPE === this.state.type}
           >
             {PointSelector.TYPE}
           </Button>
 
           <Button
             onClick={this.onChangeType}
-            active={OvalSelector.TYPE === this.state.type}
+            $active={OvalSelector.TYPE === this.state.type}
           >
             {OvalSelector.TYPE}
           </Button>
@@ -84,7 +84,7 @@ export default class Multiple extends Component {
           src={img}
           alt="Two pebbles anthropomorphized holding hands"
           annotations={this.state.annotations}
-          allowTouch={this.state.allowTouch}
+          onTouchStart={this.state.allowTouch ? undefined : (e) => e.preventDefault()}
           type={this.state.type}
           value={this.state.annotation}
           onChange={this.onChange}
