@@ -1,36 +1,31 @@
-import React from 'react'
-import styled from 'styled-components'
+import React from 'react';
+import styled from 'styled-components';
 
 const Container = styled.div`
   border: dashed 2px black;
   box-shadow: 0px 0px 1px 1px white inset;
   box-sizing: border-box;
   transition: box-shadow 0.21s ease-in-out;
-`
+`;
 
-function Rectangle (props) {
-  const { geometry } = props.annotation
-  if (!geometry) return null
+function Rectangle({ annotation, className = '', style = {}, active }) {
+  const { geometry } = annotation;
+  if (!geometry) return null;
 
   return (
     <Container
-      className={props.className}
+      className={className}
       style={{
         position: 'absolute',
         left: `${geometry.x}%`,
         top: `${geometry.y}%`,
         height: `${geometry.height}%`,
         width: `${geometry.width}%`,
-        boxShadow: props.active && '0 0 1px 1px yellow inset',
-        ...props.style
+        boxShadow: active && '0 0 1px 1px yellow inset',
+        ...style,
       }}
     />
-  )
+  );
 }
 
-Rectangle.defaultProps = {
-  className: '',
-  style: {}
-}
-
-export default Rectangle
+export default Rectangle;
