@@ -1,58 +1,49 @@
-import React, { PureComponent as Component } from 'react'
-
-const isMouseOverElement = ({ elem, e }) => {
-  const { pageY, pageX } = e
-  const { left, right, bottom, top } = elem.getBoundingClientRect()
-
-  return pageX > left && pageX < right && pageY > top && pageY < bottom
-}
-
-const isMouseHovering = (key = 'isMouseHovering') => DecoratedComponent => {
+import React, { Component } from 'react';
+const isMouseHovering = (key = 'isMouseHovering') => (DecoratedComponent) => {
   class IsMouseHovering extends Component {
     constructor() {
-      super()
-
+      super();
+      this.el = null; // Initialisation explicite.
       this.state = {
-        isHoveringOver: false
-      }
+        isHoveringOver: false,
+      };
     }
 
     componentDidMount() {
-      document.addEventListener('mousemove', this.onMouseMove)
+      document.addEventListener('mousemove', this.onMouseMove);
     }
 
     componentWillUnmount() {
-      document.removeEventListener('mousemove', this.onMouseMove)
+      document.removeEventListener('mousemove', this.onMouseMove);
     }
 
-    onMouseMove = e => {
-      const elem = this.el
+    onMouseMove = (e) => {
+      if (!this.el) {
+        return; // Évitez l'erreur si `this.el` est `null`.
+      }
 
       this.setState({
-        isHoveringOver: isMouseOverElement({ elem, e })
-      })
-    }
+        isHoveringOver: isMouseOverElement({ elem: this.el, e }),
+      });
+    };
 
     render() {
       const hocProps = {
         [key]: {
-          innerRef: el => this.el = el,
-          isHoveringOver: this.state.isHoveringOver
-        }
-      }
+          innerRef: (el) => {
+            this.el = el; // Mise à jour correcte de la référence.
+          },
+          isHoveringOver: this.state.isHoveringOver,
+        },
+      };
 
-      return (
-        <DecoratedComponent
-          {...this.props}
-          {...hocProps}
-        />
-      )
+      return <DecoratedComponent {...this.props} {...hocProps} />;
     }
   }
 
-  IsMouseHovering.displayName = `IsMouseHovering(${DecoratedComponent.displayName})`
+  IsMouseHovering.displayName = `IsMouseHovering(${DecoratedComponent.displayName})`;
 
-  return IsMouseHovering
-}
+  return IsMouseHovering;
+};
 
-export default isMouseHovering
+export default isMouseHovering;
